requirements.txt
fastapi==0.111.0
minio==7.2.7
requests==2.32.3
uvicorn==0.30.1
psycopg2-binary==2.9.9
